Fix detection of v4l stuff for 2.6.17 or older kernel
    c6e8287abe244614fa0ed4320efcc8aa85a05715
Fix compile warnings on 2.6.18 kernel
    44d88495c46d35fe3e004c3ec9363e2f6206fd36
Fix compile error with 2.6.18 kernel
    1e539a55573061939dee7958f8428456bf196fb8
Fix build with 2.6.18 kernel
    d21a94241e69fc1e622e26e2e6998861b4c8df75
pxa2xx-i2sound - PCM - disable MMAP access From: Giorgio Padrin <giorgio@mandarinlogiq.org>
    6d350a3b5603fd2dcc392911e0da24803e9e433d

+++ alsa/pci/asihpi/hpimod.c
@@ -441,8 +441,8 @@ static int __devinit adapter_probe(struct pci_dev *pci_dev,
 	for (idx = 0; idx < HPI_MAX_ADAPTER_MEM_SPACES; idx++) {
 		HPI_DEBUG_LOG4(DEBUG, "Resource %d %s %lx-%lx\n", idx,
 			       pci_dev->resource[idx].name,
-			       pci_dev->resource[idx].start,
-			       pci_dev->resource[idx].end);
+			       (unsigned long)pci_dev->resource[idx].start,
+			       (unsigned long)pci_dev->resource[idx].end);
 
 		memlen = pci_resource_len(pci_dev, idx);
 		if (memlen) {
